requirements.txt
numpy
pandas
sklearn
opencv-python
efficientnet_pytorch
transformers
albumentations
tqdm
timm
textdistance
pytorch_metric_learning
clip